requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
Flask_SocketIO==5.1.1
MarkupSafe==2.0.1
python-ldap==3.3.1
